👌 Avoid early init of scroll swizzle, use scrollViewWillBeginDragging instead of scrollViewDidScroll
